refactor(ui): moved out feedsSelector to seperate component - toggling state/visibility for feedsSelector & nodeEditor is now internal to their components - various minor tweaks
